Correct replica expiration test (#33156)
Correct the replica expiration test such that it again tests what it is
supposed to test. We achieve this by increasing the amount of data to
process.

The test ensures that replica expiration drops data that's further out
than the replica expiration. It stopped doing so after introducing (and
enabling) temporal bucketing. Temporal bucketing serves as another merge
batcher. The merge batcher first stages data in a chunker, and when
full, inserts the data in a chain.

We had so little data that it would remain in the chunker, causing us to
never reveal it in the operator/arrangement size tracking. Increasing
the data volume surfaces the data that we weren't tracking.

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>
Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>