compute: eagerly merge peek responses in Partitioned client
This commit changes how the `Partitioned` compute client handles peek
(and copy-to) responses. Instead of collecting the per-shard responses
first and then merging them once all shards have responded, we now merge
as soon as we receive a new response. This allows us to keep less state
in some cases, and ensures that the `max_result_size` checking is
actually effective in preventing OOMs.