Rework peek handling (#33260)
Start using containers in more places: use a KeyContainer to store the
literal rows we intend to lookup in the target arrangement instead of
keeping the original rows. Not intended to change performance, just good
practice, and gives us confidence that it's the right way to implement
interactions with arrangements.

---------

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>
Signed-off-by: Moritz Hoffmann <mh@materialize.com>