upsert: store errors on the heap
`DataflowError` instances are heap allocated since these values are
rarer and we don't want to be paying the cost of storing them for each
value.

The upsert operator was previosuly unpacking the general `DataflowError`
into a local `UpsertError` variant that was not boxed, leading to large
structs.

This commit re-boxes the error variant bringing the size of
`UpsertValue` down significantly.

Fixes MaterializeInc/database-issues#9567

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>