compute: check subscribe result size in Partitioned client
This commit makes the `PartitionedComputeState` enforce
`max_result_size` also for subscribe responses. This is a bit
inefficient because we have to iterate over all rows and sum up their
sizes. Using `RowCollection` would avoid this, but that's not
straightforward as they don't support timestamps currently.

We think that the inefficient approach is fine as a stop-gap protection
against ooming environmentd. Most subscribe results will be small, plus
we are already doing the same iter-and-sum result size calculation on
the cluster side and haven't noticed performance issues with that yet.