adapter: exclude crash-looping clusters from 0dt caught up checks
We already have logic in the caught-up check that tries to exclude
collections whose frontier is too far behind current wall-clock time,
but that doesn't catch cases where a replica is "limping" along and
sometimes makes process. This has made the release tedious at times,
where we had to wait a long time for cutting over.

This is a more explicit approach where we straightaway ignore clusters
that only have crash-looping replicas. We don't consider and wait for
them before cutting over in a 0dt deployment.