storagectl: refactor command/response metrics
This commit refactors the storage controller's metrics about storage
commands and responses by removing the existing bytes histograms and
replacing them with bytes and message counters. The latter are labeled
by command/response type.

This removes the ability to see the distribution of message sizes, but
introduces the ability to observe message counts by command/response
type.

The new scheme is the same as what the compute controller uses, so this
commit makes things more consistent in this regard.