compute,storage: remove support for gRPC cluster protocol
This commit removes the `enable_ctp_cluster_protocols` feature flag,
making CTP the only supported transport for the compute and storage
protocols.

The code for selecting between the two protocols is removed here as
well. A bunch of dead code from the gRPC implementation is still left,
to be removed in subsequent commits.