collection_mgmt: hold lock for while sending writes
Previously, the `update_tx.send` calls could fail because we were not
holding the `{differential,append_only}_collections` lock anymore when
performing them, making it possible for the writer task to be shut down
in the time between the sender acquired the channel and the send. This
commit avoids this by making sure the corresponding lock is held across
each `send` call.