adapter: remove replica disk flag
This commit removes the `disk` flag associated with replica definitions.
Whether a replica gets a disk now solely depends on the `disk_limit`
configuration in its replica size: If the `disk_limit` is 0, no disk is
attached, otherwise the replica gets a disk.

The change is intended to simplify the logic around disk assignment,
which previously was complicated for historical reasons.

For backward compatibility reasons, the `CREATE CLUSTER REPLICA ... WITH
(DISK = false)` syntax still works for non-cc clusters, but has no
effect now, except from showing a deprecation notice.