Persist sink: only evaluate validate_part_bounds_on_write once
This solves (I think) a rare race condition during parallel workloads.
If the FlipFlags action changes persist flags in between when we decide
to use bulk writing and when we decide to compare the batch a spurious
InvalidBatchBounds error could be emitted.

Instead of reevaluating the flag at each step, we evaluate it once,
store it in a variable and use it for the lifetime of the append_batches
invocation.