adapter: introduce `swap_enabled` replica size flag
The `swap_enabled` flag is used to decide whether replicas of that size
should be configured for swap or not. A replica is configured for swap
by:

* Not attaching a `/scratch` fs.
* Ensuring memory request differs from memory limit.

Previously swap sizes would announce not wanting a scratch disk by
setting `disk_limit = 0`, but we want to be able to set the disk limit
to the swap size so it shows up in `mz_cluster_replica_sizes`, so the
new flag becomes necessary.

The new flag allows us to remove the separate `memory_request` replica
size option, avoiding further growth of the already large replica size
definitions.