adapter: reflect memory limiter config in mz_cluster_replica_sizes
For swap replicas, the memory limiter, if switched on, can reduce the
available swap space to a value less than the physical swap space. This
commit ensures the values in `mz_cluster_replica_sizes` reflect that.

Note that changes to the memory limiter config are only reflected in
`mz_cluster_replica_sizes` after a restart, even though it is possible
to change the config at runtime. Doing it this way simplifies the
implementation and I hope will be sufficient, given that we don't plan
to change the memory limiter config often. Worst case we might have to
restart environmentd, which is quick nowadays.