storage: downgrade dependency since based on dependent's upper
The storage dataflows that maintain sources and sinks don't need to hold
back their inputs to the *since* frontier of their outputs. This is
because all we need to resume ingestion/sinking is to be able to read
the inputs at `upper - 1`, so that we produce the correct diffs for
`upper` onwards.

For a long time the storage controller would propagate the overall since
frontier of a collection to its dependencies but that is overly
pessimistic. This commit changes this so that dependency read holds
follow the implied capability of the dependent, which is always `upper
- 1`.

Fixes MaterializeInc/database-issues#8897

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>