adapter: add dyncfg to force swap for cc sizes
To allow a slow rollout of swap for cc sizes via LD, this commit
introduces a `force_swap_for_cc_sizes` dyncfg. If enabled, this config
on envd startup patches the replica size definitions to ensure swap is
always enabled for cc sizes.

Once the rollout is complete the plan is to change the externally
provided cc size definitions to enable swap directly, and remove the
dyncfg again.