[EXPLAIN] clarify direction of operator fusion (#33448)
There's been some confusion about `Fused ...` in the new EXPLAIN output.
I propose changing it to say either `Fused with Parent` (for
pre-processing MFPs in `ArrangeBy`) or `Fused with Child` (for
post-processing MFPs in a variety of operators).

While I'm in there, I made `Fused with Parent` display projects, like
all of the other MFPs.

### Motivation

  * This PR fixes a recognized bug.

https://materializeinc.slack.com/archives/C0246GEHL8N/p1756217688424869?thread_ts=1756216828.219839&cid=C0246GEHL8N
https://materializeinc.slack.com/archives/C08A62E0751/p1752091227806479