[optimizer] MIR representation types (#33321)
Separates the notion of SQL types from the actual types of our
representation.

This PR introduces the new type `Repr-` types and some tests to ensure
that it behaves correctly with regard to the old `ScalarType` &co (now
renamed to have a `Sql-` prefix). It does _not_ yet alter any MIR
structures to use the `Repr` types.

### Motivation

  * This PR adds a known-desirable feature.
     https://github.com/MaterializeInc/materialize/pull/27239