Persist: clarify `reduced` semantics to refer only to spine structure
The exert function is called exclusively by the catalog force compaction
code path. In the case that there was only one spine batch with multiple
hollow batches, the exert function would incorrectly introduce empty
SpineBatches. The behavior of the `reduced` function was always using
the number of HollowBatches as a proxy for the structural shape of the
spine. With incremental compaction, there might be only one run with
a number of empty HollowBatches. This is totally legal, and not a reason
to try and force structural merges.