Switch to indexed serialization (#33512)
Switch column from sequence to indexed serialization. The difference is
that sequence stores data as a sequence of (length, data) and indexed
puts all lengths first, followed by all data. We expect this to be more
efficient if we do not read all data, or repeatedly borrow its
serialized representation.

For details, see frankmcsherry/columnar#30

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>