Remove unnecessary snapshots
Originally there was a unary version of this normalize function, and so
I used the sqlfunc macro, which generates unit tests, which generate
these snapshots via insta. Now there is no unary variant (the parser
always calls the binary variant), so the sqlfunc macro is not used, and
these tests don't exist, and so there is no snapshot produced.