adapter: consolidate catalog updated earlier in the "apply pipeline"
When applying updates that are derived from catalog changes to the
controllers we want the updates to be consolidated. Before we were doing
that right on the parsed controller updates, but we had to implement
bespoke Ord/Eq that would ignore the in-memory fields.

Now, we sort earlier in the transact/apply "pipeline", right when we
have the StateUpdates. This allows us to remove the bespoke Ord/Eq
impls, but comes with other downsides as explained in code comments.