derive replica capacity limit from replica size
To support replica sizes with different swap:memory ratios, this commit
changes the way the replica memory limiter is configured. Instead of
deriving the base limit from the memory limit and a dyncfg, the limit is
now derived from the replica size and passed to the clusterd process as
a commandline flag.

The exact logic for deriving the memory limiter limit (referred to as
"capacity" here) is as follows: If swap is enabled, and the replica size
definition specifies both a memory and a disk limit, the capacity limit
is set to memory_limit + disk_limit. If swap is not enabled, or one of
the two limits isn't set in the replica size definition, the memory
limiter remains disabled.

Changing the memory limit at runtime is still possible through the
remaining `memory_limiter_*` dyncfgs.