catalog: stop using Arbitrary in expr_cache tests
This commit changes the expr_cache tests to not use `Arbitrary` and work
with minimal hardcoded plans/exprs instead. The expression cache largely
treats the cached expressions as opaque anyway, so generating a broad
selection of them only serves to test the bincode serialization, which
isn't this test's business.

Not relying on `Arbitrary` simplifies the tests a lot, since they don't
have to work around proptest's slowness anymore. But more importantly,
it removes the last user of `DataflowDescription`'s `Arbitrary` impl,
allowing us to remove a large amount of boilerplate in a follow-up.