Hoist the sqlfunc types into BinaryFunc
This is a massive change, sorry. It absorbs all currently defined
`LazyBinaryFunc`s into `BinaryFunc`, gettin much closer to defining it
automatically. The changes are the corresponding fallout. Where we could
name a variant directly in the past, it is now a struct, so we need to
construct the `BinaryFunc` from the struct instead of just naming a
variant.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>