Persist: eagerly compact empty batches during incremental compaction
The limits nightly test seems to exhibit a pathological case wherein 
there are no updates, but the frontier moves forward frequently leading
to _lots_ of empty batches. In theory we should handle this fine, but we
don't (I'm actively investigating this). Anyways, this seems like a good
balance to strike between the old compaction behavior and incremental
compaction, while avoiding some of the empty batch downsides.