compute: propagate monotonicity through indexes
Monotonicity information was previously being lost when the monotonic
object was read through an arrangement. The reason was that the
`PhysicallyMonotonic` LIR transform would recurse to the leaf of of an
`ArrangeBy` node, find the `Get` of the underlying collection, and check
the monotonicy there. The result is then propagated upwards.

The correctness of this transform relies on the `monotonic_ids`
configuration to include all monotonic leaf ids. The problem is that in
the case of indexes we were populating `monotonic_ids` with the id of
the index itself, which is not a leaf, and not the underlying
collection.

The solution is to instead populate the list of monotonic ids with the
id of the object the index is on.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>
Co-authored-by: Gabor Gevay <gabor@materialize.com>