adapter: default disk limit to MAX in mz_cluster_replica_sizes
Now that self-managed Materialize is a thing, users can provide their
own replica sizes, so it is no longer the case that sizes without
resource limits specified can only appear in dev environments. This is
especially relevant since self-managed users will have to set their disk
limit to either `None` or 0 for swap replica sizes, to avoid engaging
the memory limiter.

In `mz_cluster_replica_sizes`, we only have non-nullable columns for the
resource limits, so we have to pick some value. `u64::MAX` is the most
sensible one since it makes it obvious that it is just a placeholder. We
were already using it for the CPU and the memory limits, but for the
disk limit we were using 1GiB for some reason. This commit changes the
placeholder disk limit to `u64::MAX` as well.