clusterd: expose heap usage and limit
This commit extends the usage metrics provided by clusterd's
`/api/usage-metrics` endpoint with the memory usage, as well as the
effective heap (memory+swap) limit.

As for the existing usage collection, we mainly focus on Linux support,
for macOS we only provide the required stubs to make sure the code
compiles and run.