storage: fix handling of single replica ingestion
This is another instance of code that was relying on there being a
source export whose global id was the same as the ingestion. The code in
question is checking to see if the particular global id belongs to a
single-replica ingestion (e.g pg/mysql/sqlserver) but only checks if
that global id is in the list of seen source exports and sinks.

Under the new world the ingestion id does not necessarily exist as a
subsource and so we must also check in the list of ingestions.