Catalog: add migration for mock_authentication_nonce setting
The
[SASL/SCRAM](https://github.com/MaterializeInc/materialize/pull/33468)
PR introduces the need for some stable, cluster wide, cryptographically
random key material. We use this material to be able to present
deterministic challenges for even users that don't exist to guard
against enumeration attacks. 

However that PR made a bad assumption that the initialize step of
catalog opening would always add this new key. But old versions that
have already been initialized wouldn't have it! This PR add code to
generate it for old versions