adapter: use persist schema evolution for builtin migration
This commit starts using persist schema evolution for builtin item
migration. For each builtin storage collection to migrate, we first
attempt to perform a schema evolution. The attempt may fail, if persist
considers the new schema not compatible with the old one, in which case
the existing shard replacement approach is performed as a fallback.