adapter: set `mz_cluster{_replica}s.disk` to false with swap
This commit changes the meaning of the existing `disk` field in
`mz_clusters` and `mz_cluster_replicas` from "whether this
cluster/replica has a disk or swap" to "whether this cluster/replica has
a disk". I.e. the change is that the field shows `false` for
swap-enabled clusters/replicas.

The motivation is two-fold:
* Whether a replica gets swap or not depends on the node it gets
  scheduled on. Pretending that swap-enabled replicas always get swap
  would be lying.
* More importantly, the console needs a way to know which clusters still
  use the old way of spilling to disk, to adjust its metrics UI
  accordingly. This change provides that way.