Turn normalize into unary function
Normalize is a special function in that its second argument is a keyword
instead of an expression. Move from a runtime string dispatch to
determining the correct implementation at plan time. A strangeness this
introduces is that there are a bunch of functions `normalize_*` that do
not directly correspond to a function the user can type.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>