clusterd: usage_metrics: fix discovery of cgroup path
The usage_metrics code assumed that a clusterd process would only be
running within a single cgroups v2 hierarchy. But as it turns out,
kubernetes may run a process under a hybrid configuration with both v1
and v2 cgroups hierarchies.

```
$ cat /proc/7/cgroup
1:name=systemd:/
0::/
```

The previous parsing code would end up with an invalid path in this
case. This is fixed in this commit by looking specifically for the
cgroups v2 hierarchy.