Revert "cluster-spec-sheet: serializable isolation (#33807)" (#33850)
This reverts commit e291ae91db41a8299d41b40cf0219c16d6f34dcf.

Running in serializable isolation causes unexpected behavior with
materialized views: We can read from a materialized view in serializable
isolation even when the replica writing the materialized view falls
behind. This is usually not what we want to measure, so as a first step
restore the old behavior.

In the future, we can implement an approach where we specify the
isolation level for each measurement.