coord: don't take the DDL lock when sequencing ALTER SINK
Changing a sink source relation with `ALTER SINK ... SET FROM ...`
requires waiting for the old relation to catch up to the new relation.
If the old relation is stuck, the `ALTER SINK` can hang forever. We must
make sure not to disable all DDL when this happens.

The failure mode is similar to that of `ALTER CLUSTER`, and we can
employ the same workaround, namely avoiding to take the DDL lock for
`ALTER SINK`. This means in `ALTER SINK` sequencing we now have to
assume that the catalog changed while we waited for the source frontier
to advance, so we need to set up additional checking to ensure we abort
if the catalog changed in an incompatible way.