persist: register schema at write time, not writer open time
This commit changes persist clients to defer calling `register_schema`
on a shard from write handle creation time to the time of the first
append operation. The rationale is that we don't need to enforce a
schema if we are not attempting to write to a shard.

The plan is for 0dt upgrades to make good use of the new, more lenient
behavior. Read-only environments can open write handles with evolved
schemas without having to durably write down the new schemas. This will
allow us to back out of version upgrades without the risk of permanently
poisoning the persist state for lower versions.