github: fix Slack Label Notifications workflow
The Slack Label Notifications workflow requires access to the
`SLACK_TOKEN` secret, but for security reasons, workflows triggered from
forked repositories don't get access to the secrets.

The fix is to use the `pull_request_target` trigger, which bypasses the
restriction by letting the workflow run in the context of the PR base,
instead of the context of the merge commit.