ore/netio: remove some dead code
This PR removes some socket code that became dead code when we removed
the protobuf cluster protocol. Not a lot of code, but it tends to get
into the way of simulation testing tools, like turmoil and madsim, which
attempt to be tokio drop-ins but are often missing some pieces.