Lighten some catalog references
We had a bunch of places with a `&mut Catalog` or an `Arc<Catalog>`
where a simple `&Catalog` is enough. `&Catalog` is better
performance-wise than `&mut Catalog`, because creating the latter
(with calling `catalog_mut`) clones the whole catalog if there are
any other Arcs to it. Similarly, `&Catalog` is also better than an
`Arc<Catalog>`, because the latter's existence can cause other
`catalog_mut` calls to make a clone.