adapter: builtin schema migration
This commit introduces the new "builtin schema migration" mechanism. It
fulfills the same purpose as the existing "builtin item migration" but
avoids its known issues around concurrency and rollback-ability, as well
has having a (hopefully) more precise name.

In contrast to the old mechanism, the new one explicitly defines a list
of migrations to perform at specific versions, making it less likely
that we'll accidentally migrate (and thereby truncate) builtin
collections, and providing a way to choose a migration mechanism.

This commit only adds the new implementation. Subsequent commits deal
with wiring it up, populating the `MIGRATIONS` list, and deleting the
old implementation.