catalog: introduce a migration_version
This commit introduces a new `migration_version` setting stored in the
durable catalog. It can be used by migrations, like the upcoming builtin
schema migrations, to identify the source version at which the migration
starts (the target version is always the current binary version).

Currently, the `migration_version` is only used for informational
logging ("migrating from X to Y"). Note that this logging was previously
broken because it used the `catalog_content_version`, which is already
updated to the binary version by the time migrations run.

Also note that this doesn't add a migration to initialize the
`migration_version` for existing catalogs. It seems like it should be
possible to add one as a protobuf migration, that simply sets
`migration_version = catalog_content_version` (protobuf migrations do
run before `catalog_content_version` gets bumped). However, protobuf
migrations are not allowed to change any of the stored settings, because
those settings are already accessible before those migrations run. So
instead, this PR accepts that the `migration_version` will be initially
unset for old catalogs and that code using it will need to work around
that. Here we simply accept that the informational logging will be
incorrect for one (more) release.