Include the version number in the debug string
A downside of repurposing the applier_version as the actual state
version is that we can no longer determine what the actual version of
the code was that made the change withought a serious investigation.
Adding it to this string means it will be included in various places in
state.