catalog: ad-hoc migration for migration_version
This commit adds an ad-hoc migration to initialize the
`migration_version` setting for existing catalogs. Having this gets us
into a sane state where an absent `migration_version` always means the
catalog is new. Without this, it could also mean that the catalog was
previously at a version before `migration_version` was introduced, with
no easy way to differentiate the two, causing confusion in components
relying on the `migration_version`.

Having the migration in `mz-catalog` is awkward because the durable
catalog doesn't know about this setting, it's an adapter-level thing.
However, this is the only place where we can do the migration since at
the point where the opened catalog becomes accessible to adapter, the
`catalog_content_version` has already been bumped and we have lost
knowledge of the previous version. Luckily we won't have to keep this
migration around for long.