adapter: key migration shard by deploy generation too
This commit adds the deploy generation to the `Key` used in the
migration shard. This is necessary because some of our tests simulate
0dt upgrades by bumping the deploy generation without actually bumping
the version, and those tests get confused when the deploy generation
isn't in the key.

Note that this requires changing the serialization format of the
migration shard key. We choose JSON, to make the format easier to modify
in the future. We still need to be able to parse the older format so we
can clean up old entries.