ore: ensure Sentry transport never shuts down
Previously, the `mz_ore::tracing::configure` would initialize the Sentry
transport and then return a `TracingGuard` that needed to be kept alive
to prevent the Sentry transport from shutting down. Callers would
usually keep the guard until the end of the current function scope. This
caused us to miss reporting panics when a fatal error was bubbled up to
the process' `main` before becoming a panic, and the guard was dropped
in the process.

The fix taken here is to immediately forget the guard, to make sure the
Sentry transport remains intact for the process lifetime. If there is
ever a need to shut down the Sentry transport again we'll need to
reconsider, but we can keep things simple for now.