adapter: fix potential panic in builtin schema migration
The builtin schema migration code had a wrong assumption, namely that
all processes at the same version will want to replace the same shards,
and therefore write the same set of entries to the migration shard. This
is not correct! The set of shards to replace depends not only on the
target version, but also the source version. Thus, the following can
happen:

* Environment is at (source) version N. There are two `Replacement`
  migrations:
    { object: s1, version: N+1 } and
    { object: s2, version: N+2 }
* Process at version N+2 starts in read-only mode. It looks at the
  migrations, determines that it should replace both s1 and s2, and
  writes the two to the migration shard.
* Process at version N+1 starts in leader mode. It looks at the
  migrations, determines that it should replace s1, and writes that to
  the migration shard. Then it bumps the source version to N+1.
* Another process at version N+2 starts. It looks at the migrations,
  determines that it should replace s2, reads the migration shard, and
  panics because there it finds (s1, s2).

The fix is to not panic when the migration shard contents don't match,
but instead to perform a reconciliation by ignoring all the entries for
objects we don't want to replace, and appending new entries for objects
not yet in the migration shard.