adapter: pull `txn` out of the core builtin schema migration logic
This commit reorganizes the builtin schema migration code by pulling all
`Transaction` handling into the top-level `run` function. `Migration`
becomes unaware of the existing of the `Transaction` and instead returns
a `MigrationRunResult` that informs the caller about the changes that
need to be applied.

The reason for doing this is that it makes it easier to test the core
migration code. In tests it's difficult to construct a `Transaction`
object, and with this change we don't have to do that anymore, as long
as we're invoking `Migration::run` directly.