Per-object allow writes (#34039)
Change the infrastructure to allow writes to granular per object,
instead of per compute replica. While this isn't strictly needed at the
moment, it'll serve a purpose for altering materialized views where we
need to control whether a materialized view is permitted to write
independently of whether a replica is permitted to write.

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>