Reduce dependencies of catalog-protos
The mz-catalog-protos crate has a large number of dependencies that aren't
strictly needed. Sepcifically, it transitively depends on timely and the
likes, which is not ideal for a crate that is supposed to maintain the
catalog. This change aims at extracting types that have a representation
in the catalog into a separate crate that does not have any odd dependency.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>