ctp: make channels unbounded
This commit makes the channels used by the CTP implementation to connect
the send and recv tasks with the `Connection` into unbounded channels.

It was nice to have them bounded for backpressure, but unfortunately
this leads to a deadlock on the server side. Specifically, if the
connection is configured without a timeout and the send task isn't able
to send messages for whatever reason, its channel can fill up. This in
turn can indefinitely block the server's connection task. Reconnecting
doesn't help in this case, as canceling the connection task requires
that task's cooperation (for cancel safety reasons).

Having unbounded channels introduces an opportunity for unbounded memory
growth, but that seems to be the least of all evils. We use unbounded in
a bunch of places already and so far we've not had many issues because
of that, which is reassuring.

The obvious alternative would be to always configure a send timeout.
This is tricky because of self-managed though: We can't rely on users
always choosing reasonable timeouts and cluster connection issues are
notoriously hard to debug, so we want to make cluster connections as
foolproof as possible. We could set a maximum timeout of say, one
minute, but even then a misconfigured timeout would cause cluster
reconnects to take very long, degrading UX for non-obvious reasons.