persist: improve ApplyMergeResult specificity
In the move to incremental compaction we lost some specificity in the
ApplyMergeResult metric since everything is applied via
`perform_subset_replacement`. Instead of unconditionally returning
`ApplyMergeResult::AppliedSubset` we check if the range we are replacing
covers the whole batch, and if so return `AppliedExact` instead. This
should return our metric fidelity to their earlier levels.