storage: support "view" collections for other data sources
This commit changes the storage controller's data structures to support
specifying a "view" storage collection, i.e., a storage collection that
points to the same persist shard as another, for other data sources than
`Table`. The intent is specifically to get support for
`DataSource::Other`, in preparation for `ALTER MATERIALIZED VIEW`.

Rather than adding a `primary` field to the `DataSource::Other` variant,
this commit instead moves the `primary` field out of `DataSource::Table`
and into `CollectionDescription`. This reflects the fact that the
"primary" concept is independent of the data source, even though it's
currently only used for tables. It also removes some visual clutter.