adapter: remove `apply_item_update` optimization
When applying an item `StateUpdate`, we had an optimization that would
skip parsing the `create_sql` if it hadn't change, and reuse the old
`CatalogEntry` instead. This produces the wrong outcome if the
`extra_versions` change without the SQL changing: Any version updates
are not reflected in the new `CatalogEntry`.

To fix this, this commit removes the optimization, under the assumption
that DDL is sufficiently rare that we can live without it. If that turns
out not the be the case, we can keep the optimization and refine the
conditions under which SQL parsing is performed, though that seems
brittle.