adapter,sql: move sink FROM check into planning
This commit moves the check that ensures that a sink's FROM collection
is a storage collection from sequencing into planning. Generally, it's
good for errors to be caught as early as possible, to avoid unnecessary
work.

This also fixes the bug that `ALTER SINK ... SET FROM` would skip
checking the FROM collection's type and then panic during sequencing.
That's because sink planning is shared between `CREATE SINK` and `ALTER
SINK`, whereas sequencing is not.