test: remove non-essential statement that results in flakiness
It is legal for `ALTER SINK` to happen at a timestamp that is before the
two `INSERT` statements preceding it if the table hasn't yet compacted
and the sink upper is sufficiently held back. When this happens the test
fails, introducing flakiness.

I can't see a good reason for having those ignored inserts in there, we
only need to verify that alter sink works when the new from collection
has a higher global id than the sink itself, which we do.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>