envd: fix a flaky unit test
This commit unflakes a unit test (previously `test_github_12951`, now
renamed to match the current database-issues numbers).

The test would start a subscribe, then drop the subscribe's cluster,
then cancel the subscribe. This was necessary because adapter used to
not abort subscribes when their cluster was dropped. Without canceling
the subscribe, it would simply hang forever.

We since fixed this behavior, so now the subscribe is canceled when its
cluster is dropped, making the cancellation unnecessary. What's worse,
the cancellation now also makes the test flaky. The postgres client
issues the cancellation by spawning a new thread, creating a new
database connection, and issuing a cancel request on that connection.
It does not wait for the cancellation to be successful. So in the test,
an arbitrary query following the subscribe could be canceled, causing
the test to fail.

The fix is to remove the cancellation request, as it's not needed
anymore.