test: fix busy replication slot test
The test's intention was to spawn a reader of the replication and ensure
that Materialize can successfully reclaim it by killing the appropriate
server process.

However, the WAL reader that we spawned also advanced the slot forward
without Materialize having seen and committed that section. To avoid
this we pass an extremely long fsync-interval parameter that will result
in the slot being busy but not advanced.

Fixes MaterializeInc/database-issues#8447

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>